refactor(F0Graph): modularize into hooks and adopt canonical layout
Split the 1843-line F0Graph.tsx into a thin ReactFlowProvider shell plus a
F0GraphView that composes focused hooks (useExpandState, useSelectionFocus,
useGraphViewport, useGraphKeyboard, useGraphRenderModel), with pure helpers in
utils.ts and shared values in constants.ts. Move the public subcomponents under
components/ to match F0Accordion/F0AvatarList. Add a generic, configurable
F0GraphSkeleton owned by the component and consume it from the OneDataCollection
graph visualization. Public API is unchanged.
