+++ packages/react/src/patterns/F0Graph/components/F0GraphEdge/__stories__/F0GraphEdge.stories.tsx
@@ -2,9 +2,9 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 
 import { withSnapshot } from "@/lib/storybook-utils/parameters"
 
-import type { GraphEdge, GraphNode } from "../../types"
+import type { GraphEdge, GraphNode } from "../../../types"
 
-import { F0Graph, type F0GraphNodeRenderContext } from "../../F0Graph"
+import { F0Graph, type F0GraphNodeRenderContext } from "../../../F0Graph"
 import { F0GraphNode } from "../../F0GraphNode"
 import { F0GraphEdge } from "../F0GraphEdge"
