+++ packages/react/src/patterns/F0Graph/components/F0GraphNode/__stories__/F0GraphNode.stories.tsx
