+++ packages/react/src/patterns/F0Graph/components/F0GraphNode/index.tsx
