+++ packages/react/src/patterns/F0Graph/components/F0GraphNode/variants.ts
