+++ packages/react/src/patterns/F0Graph/components/F0GraphView/index.ts
@@ -1,0 +1,1 @@
+export { F0GraphView } from "./F0GraphView"
