+++ packages/react/src/patterns/F0Graph/index.tsx
@@ -1,5 +1,8 @@
 /** Interactive tree/graph visualization with zoom levels, expand/collapse, and selection. */
 export { F0Graph } from "./F0Graph"
+/** Loading placeholder mirroring the org-tree shape (root + connecting bus + children). */
+export { F0GraphSkeleton } from "./F0GraphSkeleton"
+export type { F0GraphSkeletonProps } from "./F0GraphSkeleton"
 /** Props for configuring F0Graph — data, rendering, zoom, selection, layout, and customization. */
 export type { F0GraphProps } from "./F0Graph"
 /** Render context passed to the renderNode callback. */
@@ -23,13 +26,13 @@ export type {
 
 // Sub-components
 /** Zoom/pan/fit controls toolbar. */
-export * from "./F0GraphControls"
+export * from "./components/F0GraphControls"
 /** Edge renderer with variant styles (default, hover, highlighted, dimmed). */
-export * from "./F0GraphEdge"
+export * from "./components/F0GraphEdge"
 /** Collapse/expand pill shown below parent nodes. */
-export * from "./F0GraphExpander"
+export * from "./components/F0GraphExpander"
 /** Card wrapper for nodes with avatar, title, subtitle, metadata, and actions slots. */
-export * from "./F0GraphNode"
+export * from "./components/F0GraphNode"
 
 // Hooks
 /** Focus context hook for reading/controlling roving tabindex focus within the graph. */
